requirements.txt
# Poseek Annotation Tool Dependencies
# Installation: pip install -r requirements.txt

PySide6>=6.5.0  # GUI framework
PyYAML>=6.0     # YAML configuration
pandas>=1.5.0   # CSV data processing
Pillow>=9.0.0   # Image processing
